refactor: move `handle_traces` into `cast`
