+++ crates/cast/Cargo.toml
@@ -26,6 +26,7 @@ foundry-block-explorers.workspace = true
 foundry-common.workspace = true
 foundry-compilers.workspace = true
 foundry-config.workspace = true
+foundry-debugger.workspace = true
 foundry-evm-core.workspace = true
 foundry-evm.workspace = true
 foundry-wallets.workspace = true
