+++ crates/cli/Cargo.toml
@@ -17,7 +17,6 @@ forge-fmt.workspace = true
 foundry-block-explorers.workspace = true
 foundry-common.workspace = true
 foundry-config.workspace = true
-foundry-debugger.workspace = true
 foundry-evm.workspace = true
 foundry-wallets.workspace = true
